Add worlds directory during image build time
I discovered that in the latest version of Docker that if a directory
does not exist in the container that is referenced as a volume within
the image, then the directory is created in the container as root,
regardless of what the runtime user is set to.

This ensures that the world directory is created, and that the
permission on it is set correctly.